requirements.txt
Django>=3.2.4, <3.3 #4.2.13, <4.3
Djangorestframework>=3.12.4, <3.13 #3.15.1,<3.16
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0